Client-side token validation
Since the JWT is available client side, we can just read
it and check it's presence and validity. It's signed, not encrypted.

This saves us a round-trip to the server.
We can later improve upon this by storing it in a cookie
instead of sessionStorage, since then it will also work
for the initial client-side render.

We abstracted this logic into a custom hook called useAuthGuard.