Fix razduzenje
First of all, move the submit functionality into a function,
long lines are not readable.

Using a page refresh in an SPA is bad practice.
It also breaks sometimes, because the submit request will be terminated
in some cases due to the refresh. On firefox this shows up as a network
error for me.
Instead of a refresh, we just modify the data by setting new state.

Note how we use array.filter() to return a new array instead of
modifying the existing one, because setState expects a new reference.