Fix zaduzenje
Moved it to a seperate function for readability.
Error handling is done at the top using an early return.

The problematic part was pushing the route after submit without waiting
for the submit to finish. Sometimes it would try to show the Razduzenje
page before submit was finished. Simply fixed by adding an await.